Feature update
Improve messaging
Change messages
Change bot behaviour

The bot no longer automatically deletes its own comments. This change was made because people will not expect that a comment will automatically be deleted, and they lose the chance to give feedback.

The bot now scans its inbox for deletion requests made via a button on the comment. Only the OP can delete the message, but only if there are no replies (as this would result in a visible “[deleted]” comment corpse).